[feat/#73] dto 중복 수정
