+++ app/src/main/java/com/poti/android/data/remote/dto/response/user/ProfileResponseDto.kt
@@ -24,13 +24,3 @@ data class ProfileResponseDto(
     @SerialName("recruitSummary")
     val recruitSummary: RecruitSummaryDto,
 )
-
-@Serializable
-data class RecruitSummaryDto(
-    @SerialName("total")
-    val total: Int,
-    @SerialName("inProgress")
-    val inProgress: Int,
-    @SerialName("completed")
-    val completed: Int,
-)
